Add a page for RPM Python API
We agreed that having examples as scripts in the RPM git repository
would be ideal (syntax highlighting, the possibility to run them, etc)
but for better discoverability, I would like to mention them even from
the RPM website. This way they may be easier to find for people
searching on Google or the website itself.